pyproject.toml
[project]
name = "langgraph-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "gradio>=5.46.1",
    "httpx>=0.28.1",
    "ipywidgets>=8.1.5",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.20",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "langsmith>=0.3.18",
    "lxml>=5.3.1",
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "psutil>=7.0.0",
    "pypdf>=5.4.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "semantic-kernel>=1.25.0",
    "sendgrid>=6.11.0",
    "setuptools>=78.1.0",
    "smithery>=0.1.0",
    "speedtest-cli>=2.1.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

